Implementing Media Library module
--HG--
branch : 1.x

+++ src/Orchard.Web/Modules/Orchard.MediaProcessing/Media/StorageProviderExtensions.cs
@@ -2,27 +2,20 @@
 
 namespace Orchard.MediaProcessing.Media {
     public static class StorageProviderExtensions {
-        public static bool FileExists(this IStorageProvider storageProvider, string path) {
-            try {
-                storageProvider.GetFile(path);
-                return true;
-            }
-            catch {}
-            return false;
-        }
-
         public static void TryDeleteFolder(this IStorageProvider storageProvider, string path) {
             try {
-                storageProvider.DeleteFolder(path);
+                if (storageProvider.FolderExists(path)) {
+                    storageProvider.DeleteFolder(path);
+                }
             }
             catch {}
         }
 
         public static IStorageFile OpenOrCreate(this IStorageProvider storageProvider, string path) {
-            try {
+            if (!storageProvider.FileExists(path)) {
                 return storageProvider.CreateFile(path);
             }
-            catch {}
+
             return storageProvider.GetFile(path);
         }
     }
